[autopatching] Patch -> Patcher
